requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0

# PySpark dependencies
pyspark>=3.4.0
py4j>=0.10.9

# Optional: For better performance
findspark>=2.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
